tmux: Disable escape-time
When using tmux with vscode's terminal, this option causes a stream of
commands to be sent straight to tmux when it first starts, usually
vscode sending a `cd` to the terminal so it gets opened at the root of
the project.

Disabling escape-time here stops this stream of garbage, and I see no
reason to keep it in the first instance.