nvim: Refactor my config
Decided a rework was in order because it was getting a bit hairy.

* Replaced the existing git submodules plugin arrangement with lazy.nvim.
Git submodules is flawed in many ways, chiefly in how complicated it can be.
Did I commit the library bump? Did I use the right version of that library?
Leaving it to lazy.nvim makes it a lot easier.

* Reworked using the kickstart.nvim pattern, while keeping the modular layout
for various bits of my configuration. Otherwise behaves as before.