alacritty: Retire Alacritty config
I'm using either the default Gnome Terminal (which is a bit old
at this point but if I'm on a machine which has it I can use it) or
(more recently) using Ghostty. I don't want to maintain a third
set of configurations for a terminal emulator, so I'm retiring
Alacritty.